Cap time delta to prevent huge movements after process sleeps
When the tab is hidden the browser sleeps the process and we get
exponentially backed-off animation frames. When we awake the time delta
is huge and causes the positions to be massively incremented, thus
increasing the energy in the system and ending in chaos.

Better to assume that we actually slept and just move for a maximum time
delta of 100ms. On my machine frames are typically ~20ms.